requirements.txt
streamlit
google-generativeai
python-dotenv
langchain
pypdf
chromadb
faiss-cpu